Fixed issue where the fast forward Wilys boat code was breaking the game in other ways. The nop there is unwritten after leaving the area.

+++ source/MMLAP/Models/AddressData.cs
@@ -1,12 +1,12 @@
 ﻿namespace MMLAP.Models
 {
     public class AddressData(
-        ulong address,
+        uint address,
         int? bitNumber = null,
         int? byteLength = null
         )
     {
-        public ulong Address { get; set; } = address;
+        public uint Address { get; set; } = address;
         public int? BitNumber { get; set; } = bitNumber;
         public int? ByteLength { get; set; } = byteLength;
     }
